35/requirement.txt
opencv-python
numpy
matplotlib
copy
Pillow
mediapipe
arabic_reshaper
bidi.algorithm
PySide6